[python] Geneformer Dataset builder (#762)
Adds an experimental GeneformerTokenizer to generate a Hugging Face Dataset with Geneformer token sequences derived from Census query results. This is factored with an abstract CellDatasetBuilder that captures the general pattern of processing Census query results into a cell Dataset, and a concrete subclass implementing the Geneformer tokenizer particulars within that pattern.