Make default soma context available for customization (#895)
- Add get_default_soma_context() method to obtain a TileDBSOMAContext with
reasonable defaults for Census. This allows user to customize the context
without having to also specify the default settings that are usually always
needed.
- Changed the default context to set "vfs.s3.no_sign_request" to true, since
this is the more frequent case. This avoids users having to set it when further
customizing other soma settings in a default soma context.
- Changed the default context to set "vfs.s3.region" to "us-west-2". This
avoids users having to set it when further customizing other soma settings in
a default soma context.
- Updated tests that create soma contexts use get_default_soma_context()
- Docs for default soma context usage