Balance distributed partition size more evenly
Generate the partitions by splitting on obs_joinids instead of on soma
chunks.

Previously, when the number of soma chunks was not evenly divisble by
the partition count ("world size"), partition sizes could become drastically
imbalanced.

Now, the worst case imbalance is that any two partitions will only
differ in size by 1 row. This is due to the fact the numpy.array_split()
distributed imbalances across the splits that it produces.