vector indexing pipeline for Census embeddings (#1122)
Adds a Docker+WDL pipeline to build the [TileDB-Vector-Search](https://github.com/TileDB-Inc/TileDB-Vector-Search) index for each Census embedding. (Excluding NMF for now due to some minor technical issues that will be easier to deal with after anticipated feature additions to TileDB-Vector-Search.)

The pipeline launches a powerful instance to build the index for each embedding, and processes the embeddings in parallel. The longest takes 9-10 hours to build (UCE, which has the highest dimensionality). The output is an S3 folder which can be copied into the destination bucket. 

Closes #1112